extract CustomDogfoodNavbarItem
